+++ bin/transform_to_xml/parser.py
@@ -3,17 +3,19 @@ Run this from the root folder:
 $ python bin/transform_to_xml/parser.py
 """
 
-import os, json, re
+import os
+import json
+import re
 from xml.etree.ElementTree import Element, SubElement
 import xml.etree.ElementTree as ET
 from googletrans import Translator
 
 
-def translate(tags, cat, desc, title, fromLanguage, toLanguage):
-    translator = Translator(service_urls=['translate.google.com','translate.google.com.hk','translate.google.com.tw'])
-    # Bulk process
-    _tags, _cat, _desc, _title = translator.translate([tags, cat, desc, title], dest=toLanguage, src=fromLanguage)
-    return _tags.text, _cat.text, _desc.text, _title.text
+def translate_all(data, fromLanguage, toLanguage):
+    translator = Translator(service_urls=['translate.google.com',
+                                          'translate.google.com.hk',
+                                          'translate.google.com.tw'])
+    return translator.translate(data, dest=toLanguage, src=fromLanguage)
 
 
 def gen_files(subtitle_code=None):
@@ -21,19 +23,22 @@ def gen_files(subtitle_code=None):
     TODO: there's a lot of hard cord part here ! Might be worth fixing
     :param subtitle_code: select from the list of ['ar', 'en', 'zh-CN', 'zh-TW']
     """
-    assert subtitle_code is not None, "Please supply a language code: ['ar', 'en', 'zh-CN', 'zh-TW']"
+    if subtitle_code not in ['ar', 'en', 'zh-CN', 'zh-TW']:
+        raise Exception(
+            "Please supply a language code: ['ar', 'en', 'zh-CN', 'zh-TW']")
+
+    subtitles_dir = os.path.join('subtitles', 'tedEd')
     cwd = os.getcwd()
-    folder_names = os.listdir('./subtitles/tedDirector')
+    folder_names = os.listdir(os.path.join(cwd, subtitles_dir))
     for folder in folder_names:  # EACH VIDEO!
-        subtitles = os.listdir(cwd + '/subtitles/tedDirector/' + folder)
+        subtitles = os.listdir(os.path.join(cwd, subtitles_dir, folder))
         for subtitle in subtitles:  # EACH SUBTITLE IN THE VIDEO
             lang = subtitle.split('.', 2)[-2]
             if lang == subtitle_code:
-                filepath = cwd + '/subtitles/tedDirector/' + folder + '/' + subtitle
-                yield filepath
+                yield os.path.join(cwd, subtitles_dir, folder, subtitle)
 
 
-def parse_simple(filepath, language):
+def parse_simple(filepath, language, meta_data):
     """
     Following the convention of Trectext, the following tags are used to supply additional metadata for the video:
     1. <DOCNO>      ::  Corresponds to the video id
@@ -43,147 +48,109 @@ def parse_simple(filepath, language):
     5. <leadpara>   ::  the description of the video as in youtube.
     6. <head>       ::  the category the youtube video belongs to
     """
-    start = False  # Flag to control when to start including the text.
     _et = Element('DOC')
 
-    video_id = str(filepath).split('/')[-2]
+    video_id = os.path.basename(os.path.dirname(subtitle_file))
     _id = SubElement(_et, 'DOCNO')
     _id.text = video_id.encode('utf-8')
     link = 'https://www.youtube.com/watch?v=' + video_id
     _link = SubElement(_et, 'LINK')  # this will not be read by trec parser.
     _link.text = link.encode('utf-8')
 
-    # add meta information
-    jsonFile = "/".join(str(filepath).split('/')[:-1]) + '/' + video_id + '.info.json'
-    with open(jsonFile) as f:
-        datastore = json.load(f, encoding='utf-8')
-        # tags = " ".join(datastore['tags'])
-        # cat = " ".join(datastore['categories'])
-        tags = datastore['tags']
-        cat = datastore['categories']
-        desc = datastore["description"].encode('unicode-escape').encode('utf-8')
-        title = datastore["title"].encode('unicode-escape')
-        if language != 'en':
-            # append fullstop so that each word/phrase is independent from each other in the MT system
-            tags = '. '.join(tags)
-            cat = '. '.join(cat)
-            tags, cat, desc, title = translate(tags, cat, desc, title, fromLanguage='en', toLanguage=language)
-
     _tags = SubElement(_et, 'HEADLINE')
-    _tags.text = tags
+    _tags.text = meta_data[0]
     _category = SubElement(_et, 'HEAD')
-    _category.text = cat
+    _category.text = meta_data[1]
     _description = SubElement(_et, 'LEADPARA')
-    _description.text = desc.replace('\\n', '')
+    _description.text = meta_data[2].replace('\\n', '')
     _title = SubElement(_et, 'TITLE')
-    _title.text = title
+    _title.text = meta_data[3]
 
     _text = SubElement(_et, 'TEXT')
     text = ""
     with open(filepath, 'r') as f:
+        start = False  # flag to control when to start including the text.
         for line in f:
             line = line.strip()
-            # print type(line) #DEBUG
             if line.startswith('Language:'):
                 _, lang = line.split(': ', 1)
                 _lang = SubElement(_et, 'LANGUAGE')
                 _lang.text = lang.encode('utf-8')
                 print(lang)
 
-            elif '-->' in line:
-                # Ignore the timestamps
+            elif '-->' in line:  # Ignore the timestamps
                 start = True
                 continue
 
             elif line and start:
-                # if language == 'en':
-                #     text += line.decode() + ' '
-                #     # print line
-                # else:
                 text += line.decode('utf-8') + ' '
 
-    # Add the text to the file:
+    # Add the text to the file
     _text.text = text
 
     return _et
 
 
-def parse_timestamp(filepath, language):
-    """
-    Parse the file given in the argument, filePath according to the VTT format.
-    dump the filepath in writeFile.
-    <SUBTITLE>
-        <VIDEOID>  </VIDEOID>
-        <LANGUAGE>  </LANGUAGE>
-        <CAPTION>
-            <TEXT>
-            <STARTTIME> </STATTIME>
-            <ENDTIME> </ENDTIME>
-        </CAPTION>
-    </SUBTITLE>
-    :param filepath: the file to write
-    :return: ET.toSting(elemenTree)
-    """
-    video_id = str(filepath).split('/')[-2]
-    timeStep = None
-    _attributes = {"start": None, "end": None}
-    attributes = _attributes
-    with open(filepath, 'r') as f:
-        _et = Element('SUBTITLE')
-        _id = SubElement(_et, 'VIDEOID')
-        _id.text = video_id.encode('utf-8')
+def get_meta_data(subtitle_file, language):
+    video_id = os.path.basename(os.path.dirname(subtitle_file))
+    json_file = os.path.join(os.path.dirname(subtitle_file),
+                             video_id + '.info.json')
+    with open(json_file) as f:
+        datastore = json.load(f, encoding='utf-8')
 
-        for line in f:
-            line = line.strip()
-            # print type(line) #DEBUG
-            if line.startswith('Language:'):
-                _, lang = line.split(': ', 1)
-                _lang = SubElement(_et, 'LANGUAGE')
-                _lang.text = lang.encode('utf-8')
-                print(lang)
+        tags = datastore['tags']
+        cat = datastore["categories"]
+        desc = datastore["description"].encode(
+            'unicode-escape').encode('utf-8')
+        title = datastore["title"].encode('unicode-escape')
 
-            elif '-->' in line:
-                start, end = line.split('-->')
-                # _startTime = SubElement(_caption, 'STARTTIME')
-                # _endTime = SubElement(_caption, 'ENDTIME')
-                attributes['start'] = start.strip().encode('utf-8')
-                attributes['end'] = end.strip().encode('utf-8')
-                # _startTime.text = start.strip()
-                # _endTime.text = end.strip()
-                _caption = SubElement(_et, 'CAPTION', attrib=attributes)
-                timeStep = _caption
+        if language != 'en':
+            # append fullstop so that each word/phrase is independent from each
+            # other in the MT system
+            tags = '. '.join(tags)
+            cat = '. '.join(cat)
 
-            elif line:
-                if timeStep is not None:
-                    # print(line)
-                    # if lang == 'en':
-                    #     timeStep.text = line
-                    # elif lang == 'ar':
-                    # timeStep.text = line.decode('utf-8') # this works, means that the line is already in unicode!
-                    timeStep.text = line.decode('utf-8')
-                else:
-                    timeStep = None
-                    continue
-    # try:
-    #     output = tostring(_et, encoding='unicode')
-    # except TypeError:
-    # output = tostring(_et)  # Return in US_ASCII!
-    # output = tostring(_et, encoding='utf-8')  # Return in unicode!
-    # print(output)
-    return _et
-
+    return [tags, cat, desc, title]
 
 if __name__ == "__main__":
     import sys
     lang = sys.argv[1]
-    assert lang in ['ar', 'en', 'zh-CN', 'zh-TW'], "Language must be either: ['ar', 'en', 'zh-CN', 'zh-TW']"
-    count = 0
+    if lang not in ['ar', 'en', 'zh-CN', 'zh-TW']:
+        raise Exception(
+            "Please supply a language code: ['ar', 'en', 'zh-CN', 'zh-TW']")
 
+    count = 0
     output = sys.argv[2]
     with open(output, 'w') as w:
-        for sub in gen_files(lang):
-            count += 1
-            elem = parse_simple(sub, lang)
+        all_meta_data = []
+        for subtitle_file in gen_files(lang):
+            all_meta_data.extend(get_meta_data(subtitle_file, lang))
+
+        meta_data_len = len(all_meta_data)
+        if lang != 'en':
+            WINDOW = 60
+            translated_meta_data = []
+            idx = 0
+            while idx < meta_data_len:
+                if idx + WINDOW < meta_data_len:
+                    chunk = all_meta_data[idx: idx + WINDOW]
+                    idx += WINDOW
+                else:
+                    chunk = all_meta_data[idx:]
+                    idx = meta_data_len
+                print('Translating %s subtitles' % (len(chunk) / 4))
+                translated_meta_data.extend([x.text for x in translate_all(chunk,
+                                                          fromLanguage='en',
+                                                          toLanguage=lang)])
+            assert meta_data_len == len(translated_meta_data)
+            all_meta_data = translated_meta_data[:]
+
+        count = 0
+        for subtitle_file in gen_files(lang):
+            meta_data = all_meta_data[4 * count: 4 * count + 4]
+            elem = parse_simple(subtitle_file, lang, meta_data)
             elementTree = ET.ElementTree(elem)
             elementTree.write(w, encoding='utf-8')
+            count += 1
+
     print(str(count) + ' files parsed')
